modules/notifications/requirements.txt
fastapi
uvicorn[standard]
resend
twilio
python-dotenv
pydantic
httpx
pytest
